Move open_archive to archive_utils module
