+++ autobuild/common.py
@@ -524,27 +524,3 @@ def has_cmd(name, subcmd: str = "help") -> bool:
     except OSError:
         return False
     return not p.returncode
-
-
-class ZstdTarFile(tarfile.TarFile):
-    def __init__(self, name, mode='r', *, level=4, zstd_dict=None, **kwargs):
-        from pyzstd import CParameter, ZstdFile
-        zstdoption = None
-        if mode != 'r' and mode != 'rb':
-           zstdoption = {CParameter.compressionLevel : level,
-                         CParameter.nbWorkers : multiprocessing.cpu_count(),
-                         CParameter.checksumFlag : 1}
-        self.zstd_file = ZstdFile(name, mode,
-                                level_or_option=zstdoption,
-                                zstd_dict=zstd_dict)
-        try:
-            super().__init__(fileobj=self.zstd_file, mode=mode, **kwargs)
-        except:
-            self.zstd_file.close()
-            raise
-
-    def close(self):
-        try:
-            super().close()
-        finally:
-            self.zstd_file.close()
